requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
python-dotenv==1.0.1
SQLAlchemy==2.0.38
PyMySQL==1.1.2
langchain==1.2.9
langchain-community==0.4.1
langchain-google-genai==4.2.0
APScheduler==3.11.2
